Quiz | Move fetch from route to moviedb service
